+++ src/lesson2/task2/Logical.kt
@@ -34,7 +34,7 @@ fun isNumberHappy(number: Int): Boolean {
  * Определить, угрожают ли они друг другу. Вернуть true, если угрожают.
  */
 fun queenThreatens(x1: Int, y1: Int, x2: Int, y2: Int): Boolean =
-        if ((x1 == x2) || (y1 == y2) || (abs((x1 - x2).toDouble()) == abs((y1 - y2).toDouble()))) true
+        if ((x1 == x2) || (y1 == y2) || (abs(x1 - x2) == abs(y1 - y2))) true
         else false
